Moved function that sets browser window size to test support file so tests will not have to duplicate the function. Added a new (empty) template file to assist writers in adding tests.
